requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==6.17.0
adafruit-circuitpython-ads1x15==2.2.10
adafruit-circuitpython-busdevice==5.1.1
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.18.0
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.17.0
Babel==2.9.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
boto3==1.17.57
botocore==1.20.112
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
contourpy==1.0.6
cookiecutter==1.7.3
cowpy==1.1.5
cycler==0.11.0
distlib==0.3.4
Django==2.2.14
django-browser-reload==1.3.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-jsonfield==1.4.1
django-tailwind==3.1.1
docutils==0.18.1
download==0.3.5
Eel==0.12.4
emoji==2.0.0
filelock==3.6.0
flake8==4.0.1
flake8-polyfill==1.0.2
Flask==2.0.2
Flask-MySQLdb==1.0.1
fontawesomefree==6.2.1
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
greenlet==1.1.2
httplib2==0.21.0
idna==3.3
itsdangerous==2.0.1
java2python==0.5.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.6.2
mccabe==0.6.1
multidict==6.0.2
mypy==0.960
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.22.3
oauthlib==3.2.2
packaging==21.3
pandas==1.4.2
parso==0.8.3
pefile==2021.9.3
pep8-naming==0.12.1
Pillow==9.3.0
platformdirs==2.5.1
poyo==0.5.0
progress==1.6
progressbar==2.5
prompt-toolkit==3.0.28
protobuf==4.21.12
psycopg==3.0.11
psycopg2==2.9.3
ptpython==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyfiglet==0.8.post1
pyflakes==2.4.0
pyftdi==0.53.3
Pygments==2.11.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pyparsing==3.0.9
pypiwin32==223
pyserial==3.5
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
pyusb==1.2.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.4.2
sentry-sdk==1.5.10
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
tkcalendar==1.6.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.8
urlopen==1.0.0
virtualenv==20.13.3
wcwidth==0.2.5
Werkzeug==2.0.2
whichcraft==0.6.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
